requirements.txt
async-generator==1.10
attrs==21.4.0
black==22.3.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
outcome==1.1.0
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.10.0
typing_extensions==4.2.0
urllib3==1.26.9
webdriver-manager==3.5.4
whichcraft==0.6.1
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
